Rename PageLayout components folder
